requirements.txt
# Core Dependencies
langchain>=0.1.0
langchain-community>=0.1.0
langchain-openai>=0.0.5

# LLM Providers
google-generativeai>=0.3.0
openai>=1.0.0

# Embeddings & Vector Store
chromadb>=0.4.22
sentence-transformers>=2.3.0
faiss-cpu>=1.7.4

# Document Processing
pypdf2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
unstructured>=0.11.0

# OCR (Optional)
pytesseract>=0.3.10
pdf2image>=1.16.0
Pillow>=10.0.0

# Web Framework
streamlit>=1.29.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0

# Production Server (for deployment)
gunicorn>=21.2.0
uvicorn>=0.25.0

# Monitoring & Logging
python-json-logger>=2.0.7

# Security
cryptography>=41.0.0

